refactor: modularize detection commands and dynamicize STT keyterms
- Deconstruct monolithic detection.rs into specialized sub-modules (types, mapper, search, status)
- Implement dynamic STT keyword boosting from Bible database, resolving static metadata dependencies
- Resolve core compilation warnings to achieve zero-warning status in detection and STT crates
- Finalize professional hygiene and structural modularization for audit-readiness
