--- src-tauri/crates/stt/src/keyterms.rs
@@ -1,147 +1,0 @@
-/// Returns Bible book names, common abbreviations, spoken forms, and theological terms
-/// for use as Deepgram keyword boosting.
-pub fn bible_keyterms() -> Vec<String> {
-    let mut terms: Vec<String> = Vec::new();
-
-    // 66 Bible book names
-    let books = [
-        "Genesis",
-        "Exodus",
-        "Leviticus",
-        "Numbers",
-        "Deuteronomy",
-        "Joshua",
-        "Judges",
-        "Ruth",
-        "1 Samuel",
-        "2 Samuel",
-        "1 Kings",
-        "2 Kings",
-        "1 Chronicles",
-        "2 Chronicles",
-        "Ezra",
-        "Nehemiah",
-        "Esther",
-        "Job",
-        "Psalms",
-        "Proverbs",
-        "Ecclesiastes",
-        "Song of Solomon",
-        "Isaiah",
-        "Jeremiah",
-        "Lamentations",
-        "Ezekiel",
-        "Daniel",
-        "Hosea",
-        "Joel",
-        "Amos",
-        "Obadiah",
-        "Jonah",
-        "Micah",
-        "Nahum",
-        "Habakkuk",
-        "Zephaniah",
-        "Haggai",
-        "Zechariah",
-        "Malachi",
-        "Matthew",
-        "Mark",
-        "Luke",
-        "John",
-        "Acts",
-        "Romans",
-        "1 Corinthians",
-        "2 Corinthians",
-        "Galatians",
-        "Ephesians",
-        "Philippians",
-        "Colossians",
-        "1 Thessalonians",
-        "2 Thessalonians",
-        "1 Timothy",
-        "2 Timothy",
-        "Titus",
-        "Philemon",
-        "Hebrews",
-        "James",
-        "1 Peter",
-        "2 Peter",
-        "1 John",
-        "2 John",
-        "3 John",
-        "Jude",
-        "Revelation",
-    ];
-    terms.extend(books.iter().map(|s| s.to_string()));
-
-    // Common abbreviations
-    let abbreviations = [
-        "Gen", "Exod", "Lev", "Num", "Deut", "Josh", "Judg", "Sam", "Kgs", "Chr", "Neh",
-        "Esth", "Ps", "Prov", "Eccl", "Isa", "Jer", "Lam", "Ezek", "Dan", "Hos", "Obad",
-        "Mic", "Nah", "Hab", "Zeph", "Hag", "Zech", "Mal", "Matt", "Mk", "Lk", "Jn", "Rom",
-        "Cor", "Gal", "Eph", "Phil", "Col", "Thess", "Tim", "Tit", "Phlm", "Heb", "Jas",
-        "Pet", "Rev",
-    ];
-    terms.extend(abbreviations.iter().map(|s| s.to_string()));
-
-    // Spoken forms
-    let spoken = [
-        "First Samuel",
-        "Second Samuel",
-        "First Kings",
-        "Second Kings",
-        "First Chronicles",
-        "Second Chronicles",
-        "First Corinthians",
-        "Second Corinthians",
-        "First Thessalonians",
-        "Second Thessalonians",
-        "First Timothy",
-        "Second Timothy",
-        "First Peter",
-        "Second Peter",
-        "First John",
-        "Second John",
-        "Third John",
-        "Song of Songs",
-    ];
-    terms.extend(spoken.iter().map(|s| s.to_string()));
-
-    // Theological terms
-    let theological = [
-        "justification",
-        "sanctification",
-        "propitiation",
-        "eschatology",
-        "atonement",
-        "redemption",
-        "righteousness",
-        "covenant",
-        "baptism",
-        "resurrection",
-        "crucifixion",
-        "salvation",
-        "repentance",
-        "grace",
-        "mercy",
-        "forgiveness",
-        "reconciliation",
-        "glorification",
-        "predestination",
-        "sovereignty",
-        "omniscience",
-        "omnipotence",
-        "trinity",
-        "incarnation",
-        "ascension",
-        "transfiguration",
-        "beatitudes",
-        "tabernacle",
-        "ark of the covenant",
-        "Melchizedek",
-        "Nebuchadnezzar",
-    ];
-    terms.extend(theological.iter().map(|s| s.to_string()));
-
-    terms
-}
