+++ src-tauri/crates/stt/src/types.rs
@@ -35,6 +35,7 @@ pub struct SttConfig {
     pub sample_rate: u32,
     pub encoding: String,
     pub language: Option<String>,
+    pub keyterms: Vec<String>,
 }
 
 impl Default for SttConfig {
@@ -45,6 +46,7 @@ impl Default for SttConfig {
             sample_rate: 16000,
             encoding: "linear16".to_string(),
             language: None,
+            keyterms: Vec::new(),
         }
     }
 }
